@dev 添加用户简历增删改查以及简历教育背景、工作经历、实习经历、在校获奖经历、在校参加活动情况等模块增查功能

+++ user-management-module/src/main/java/com/solstice/melon/domain/EducationalExperience.java
@@ -33,7 +33,7 @@ public class EducationalExperience extends BaseDomain implements Serializable {
     /**
      * 专业学制
      */
-    private Integer yearOfMajor;
+    private Double yearOfMajor;
     /**
      * 专业研究方向介绍
      */
@@ -79,11 +79,11 @@ public class EducationalExperience extends BaseDomain implements Serializable {
         this.major = major;
     }
 
-    public Integer getYearOfMajor() {
+    public Double getYearOfMajor() {
         return yearOfMajor;
     }
 
-    public void setYearOfMajor(Integer yearOfMajor) {
+    public void setYearOfMajor(Double yearOfMajor) {
         this.yearOfMajor = yearOfMajor;
     }
